requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langgraph
python-dotenv
duckduckgo-search
fpdf